Add a start-deamon command that only starts the daemon (instead of start)
start now also connects to the deamon

this allows multiple calls to start that start different scripts
