+++ src/daemon-command/restart.daemon-command.ts
@@ -1,33 +1,34 @@
-import colors from "colors";
 import { Socket } from "net";
-import { Daemon } from "src/commands/start.command";
 
+import { Daemon } from "../commands/start-daemon.command";
 import { startProcess } from "./start-process";
 
 export async function restartDaemonCommand(daemon: Daemon, socket: Socket, scriptName: string): Promise<void> {
     const { scripts, processes } = daemon;
-    const p = processes[scriptName];
-    if (!p) {
-        console.log(`${colors.bgYellow.bold.black(" dev-pm ")} unknown script name  ${scriptName}`);
+    const script = scripts.find((i) => i.name == scriptName);
+    if (!script) {
+        console.log(`unknown script name ${scriptName}`);
+        socket.write(`unknown script name ${scriptName}\n`);
         socket.end();
         return;
     }
 
-    if (!p.killed) {
-        console.log(`${colors.bgYellow.bold.black(" dev-pm ")} killing ${scriptName}`);
-        p.kill("SIGINT");
-        while (!p.killed) {
-            console.log(`${colors.bgYellow.bold.black(" dev-pm ")} waiting for killed`);
-            await new Promise((r) => setTimeout(r, 100));
+    const process = processes[scriptName];
+    if (process) {
+        if (!process.killed) {
+            console.log(`killing ${scriptName}`);
+            socket.write(`killing ${scriptName}`);
+            process.kill("SIGINT");
+            while (!process.killed) {
+                console.log(`waiting for killed`);
+                socket.write(`waiting for killed\n`);
+                await new Promise((r) => setTimeout(r, 100));
+            }
         }
     }
 
-    const script = scripts.find((i) => i.name == scriptName);
-    if (!script) {
-        console.log(`${colors.bgYellow.bold.black(" dev-pm ")} unknown script name  ${scriptName}`);
-        socket.end();
-        return;
-    }
+    console.log(`killing ${scriptName}`);
+    socket.write(`starting ${scriptName}\n`);
     startProcess(daemon, script);
     socket.end();
 }
