+++ src/daemon-command/shutdown.daemon-command.ts
@@ -1,6 +1,6 @@
 import { Socket } from "net";
-import { Daemon } from "src/commands/start.command";
 
+import { Daemon } from "../commands/start-daemon.command";
 import { shutdown } from "./shutdown";
 
 export function shutdownDaemonCommand(daemon: Daemon, socket: Socket): void {
